+++ app/uk/gov/hmrc/hec/models/hecTaxCheck/HECTaxCheckData.scala
@@ -19,7 +19,6 @@ package uk.gov.hmrc.hec.models.hecTaxCheck
 import play.api.libs.json._
 import uk.gov.hmrc.hec.models.hecTaxCheck.ApplicantDetails.{CompanyApplicantDetails, IndividualApplicantDetails}
 import uk.gov.hmrc.hec.models.hecTaxCheck.TaxDetails.{CompanyTaxDetails, IndividualTaxDetails}
-import uk.gov.hmrc.hec.models.EntityType
 
 import java.time.ZonedDateTime
