+++ app/uk/gov/hmrc/hec/models/hecTaxCheck/TaxDetails.scala
@@ -18,7 +18,6 @@ package uk.gov.hmrc.hec.models.hecTaxCheck
 
 import play.api.libs.json.{Json, OFormat}
 import uk.gov.hmrc.hec.models.ids.{CTUTR, NINO, SAUTR}
-import uk.gov.hmrc.hec.models._
 
 sealed trait TaxDetails extends Product with Serializable
