+++ app/uk/gov/hmrc/hec/models/hecTaxCheck/TaxYear.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.hec.models
+package uk.gov.hmrc.hec.models.hecTaxCheck
 
 import cats.instances.int._
 import cats.syntax.eq._
