+++ app/uk/gov/hmrc/hec/models/hecTaxCheck/YesNoAnswer.scala
@@ -19,15 +19,22 @@ package uk.gov.hmrc.hec.models.hecTaxCheck
 import ai.x.play.json.Jsonx
 import ai.x.play.json.SingletonEncoder.simpleName
 import ai.x.play.json.implicits.formatSingleton
+import cats.Eq
 import play.api.libs.json.Format
 
-sealed trait HECTaxCheckMatchStatus extends Product with Serializable
+sealed trait YesNoAnswer extends Product with Serializable
 
-object HECTaxCheckMatchStatus {
-  case object NoMatch extends HECTaxCheckMatchStatus
-  case object Match extends HECTaxCheckMatchStatus
-  case object Expired extends HECTaxCheckMatchStatus
+object YesNoAnswer {
+
+  case object Yes extends YesNoAnswer
+
+  case object No extends YesNoAnswer
+
+  implicit val eq: Eq[YesNoAnswer] = Eq.fromUniversalEquals
 
   @SuppressWarnings(Array("org.wartremover.warts.Throw", "org.wartremover.warts.Equals"))
-  implicit val format: Format[HECTaxCheckMatchStatus] = Jsonx.formatSealed[HECTaxCheckMatchStatus]
+  implicit val format: Format[YesNoAnswer] = Jsonx.formatSealed[YesNoAnswer]
+
+  val values: List[YesNoAnswer] = List(Yes, No)
+
 }
