+++ test/uk/gov/hmrc/hec/controllers/IFControllerSpec.scala
@@ -25,7 +25,7 @@ import play.api.test.Helpers._
 import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.hec.models.{hecTaxCheck, _}
 import uk.gov.hmrc.hec.models.ids.{CTUTR, GGCredId, SAUTR}
-import uk.gov.hmrc.hec.models.hecTaxCheck.{CTStatusResponse, SAStatus, SAStatusResponse}
+import uk.gov.hmrc.hec.models.hecTaxCheck.{CTStatusResponse, SAStatus, SAStatusResponse, TaxYear}
 import uk.gov.hmrc.hec.services.IFService
 import uk.gov.hmrc.hec.services.IFService.{BackendError, DataNotFoundError, IFError}
 import uk.gov.hmrc.http.HeaderCarrier
