+++ test/uk/gov/hmrc/hec/services/HecTaxCheckExtractionServiceSpec.scala
@@ -28,14 +28,14 @@ import org.scalatest.wordspec.AnyWordSpecLike
 import play.api.Configuration
 import play.api.test.Helpers._
 import uk.gov.hmrc.hec.models
+import uk.gov.hmrc.hec.models._
 import uk.gov.hmrc.hec.models.hecTaxCheck.ApplicantDetails.CompanyApplicantDetails
 import uk.gov.hmrc.hec.models.hecTaxCheck.HECTaxCheckData.CompanyHECTaxCheckData
 import uk.gov.hmrc.hec.models.hecTaxCheck.TaxDetails.CompanyTaxDetails
-import uk.gov.hmrc.hec.models.{hecTaxCheck, _}
+import uk.gov.hmrc.hec.models.hecTaxCheck._
 import uk.gov.hmrc.hec.models.ids.{CRN, CTUTR, GGCredId}
 import uk.gov.hmrc.hec.models.licence.{LicenceTimeTrading, LicenceType, LicenceValidityPeriod}
 import uk.gov.hmrc.hec.models.sdes._
-import uk.gov.hmrc.hec.models.hecTaxCheck.{CTAccountingPeriod, CTStatus, CTStatusResponse, HECTaxCheck, HECTaxCheckCode, HECTaxCheckFileBodyList, HECTaxCheckSource, LicenceDetails}
 import uk.gov.hmrc.hec.services.scheduleService.{HECTaxCheckExtractionContext, HecTaxCheckExtractionServiceImpl}
 import uk.gov.hmrc.hec.util.{TimeUtils, UUIDGenerator}
 import uk.gov.hmrc.http.HeaderCarrier
